+++ src/index.ts
@@ -22,7 +22,7 @@ export * from "./chain/tezos/TezosNodeReader";
 export * from "./chain/tezos/TezosNodeWriter";
 export * from './chain/tezos/contracts/BabylonDelegationHelper';
 export * from './chain/tezos/contracts/DexterTokenHelper';
-export * from './chain/tezos/contracts/MorleyTokenHelper';
+export * from './chain/tezos/contracts/Tzip7ReferenceTokenHelper';
 export * from './chain/tezos/contracts/TCFBakerRegistryHelper';
 
 export * from "./identity/tezos/TezosFileWallet";